moving old grating to deprecated
