--- pylum/grating_coupler/gc_clean.py
@@ -1,7 +1,0 @@
-if __name__ == "__main__":
-    import lumapi
-
-    s = lumapi.FDTD()
-    s.load("grating_coupler_2D.fsp")
-    s.switchtolayout()
-    s.save()
