+++ pylum/loadmat.py
@@ -50,7 +50,7 @@ def write_dict(d, filepath="results.json"):
 
 
 def test_loadmat():
-    from pylum.grating_coupler import sparameters
+    from pylum.deprecated.grating_coupler import sparameters
 
     matlab_file_path = (
         CONFIG["repo_path"] / "workspace" / "grating_coupler_sweep" / "results.mat"
